+++ src/datasources/alarms/query-type-handlers/list-alarms/ListAlarmsQueryHandler.ts
@@ -1,9 +1,10 @@
 import { DataFrameDTO, DataQueryRequest, DataSourceInstanceSettings, LegacyMetricFindQueryOptions, MetricFindValue } from '@grafana/data';
 import { ListAlarmsQuery } from '../../types/ListAlarms.types';
-import { AlarmsVariableQuery } from '../../types/types';
+import { AlarmsVariableQuery, QueryAlarmsRequest } from '../../types/types';
 import { AlarmsQueryHandlerCore } from '../AlarmsQueryHandlerCore';
 import { DEFAULT_QUERY_EDITOR_DESCENDING, QUERY_EDITOR_MAX_TAKE, QUERY_EDITOR_MIN_TAKE } from 'datasources/alarms/constants/AlarmsQueryEditor.constants';
 import { defaultListAlarmsQuery } from 'datasources/alarms/constants/DefaultQueries.constants';
+import { Alarm } from 'datasources/alarms/types/types';
 import { BackendSrv, getBackendSrv, getTemplateSrv, TemplateSrv } from '@grafana/runtime';
 import { User } from 'shared/types/QueryUsers.types';
 import { UsersUtils } from 'shared/users.utils';
@@ -22,7 +23,12 @@ export class ListAlarmsQueryHandler extends AlarmsQueryHandlerCore {
     this.usersUtils = new UsersUtils(this.instanceSettings, this.backendSrv);
   }
 
-  public async runQuery(query: ListAlarmsQuery, _options: DataQueryRequest): Promise<DataFrameDTO> {
+  public async runQuery(query: ListAlarmsQuery, options: DataQueryRequest): Promise<DataFrameDTO> {
+    query.filter = this.transformAlarmsQuery(options.scopedVars, query.filter);
+
+    // #AB:3449773 Map queryAlarmsData response to user-selected properties
+    await this.queryAlarmsData(query);
+
     return {
       refId: query.refId,
       name: query.refId,
@@ -60,6 +66,14 @@ export class ListAlarmsQueryHandler extends AlarmsQueryHandlerCore {
 
   private isTakeValid(take?: number): boolean {
     return take !== undefined && take >= QUERY_EDITOR_MIN_TAKE && take <= QUERY_EDITOR_MAX_TAKE;
+  }
+
+  private async queryAlarmsData(alarmsQuery: ListAlarmsQuery): Promise<Alarm[]> {
+    const alarmsRequestBody: QueryAlarmsRequest = {
+      filter: alarmsQuery.filter ?? '',
+    }
+
+    return this.queryAlarmsInBatches(alarmsRequestBody);
   }
 
   // @ts-ignore